# swagger.yaml
...
swagger: "2.0"
host: "localhost:8000"
basePath: "/api/v0.3"
info:
  title: User-Organization API
  description: An API which supports management of customers and organizations within the LEXIS portal; this version is intended for use in LEXIS Portal R1
  version: "0.3.2021-12-15"
  contact:
    email: diego@cyclops-labs.io
  license:
    name: Apache 2.0
    url: 'http://www.apache.org/licenses/LICENSE-2.0.html'
...
securityDefinitions:
  APIKeyHeader:
    type: apiKey
    in: header
    name: X-API-KEY
  APIKeyParam:
    type: apiKey
    in: query
    name: api_key
  Keycloak:
    type: oauth2
    flow:  accessCode
    authorizationUrl: 'http://localhost:8080/auth/realms/Dev/protocol/openid-connect/auth'
    tokenUrl: 'http://localhost:8080/auth/realms/Dev/protocol/openid-connect/token'
    scopes:
      lex_adm: $Deity scope, can do absoulutely anything
      lex_sup: $Deity scope with restrictions.
      org_mgr: Organization Manager
      prj_mgr: Project/Workflows Manager
      dat_mgr: Datasets Manager
      end_usr: Plain user
schemes:
  - http
  - https
security:
  - Keycloak: [lex_adm, lex_sup, org_mgr, prj_mgr, dat_mgr, end_usr]
paths:
...
  /organization:
    get:
      tags:
        - organizationManagement
      produces:
        - application/json
      summary: list organizations
      operationId: listOrganizations
      description: lists all organizations
      responses:
        '200':
          description: list of organizations
          schema:
            type: array
            items:
              $ref: "#/definitions/Organization"
        '401':
          description: Authorization error
          schema:
            $ref: "#/definitions/ErrorResponse"
        '403':
          description: Authorization error
          schema:
            $ref: "#/definitions/ErrorResponse"
        '500':
          description: unexpected error
          schema:
            $ref: "#/definitions/ErrorResponse"
      parameters:
        - in: query
          name: scope
          description: organization scope switch
          type: string
          enum:
          - ALL
          - OWN
          - ALLOWED
          default: OWN
    post:
      tags:
        - organizationManagement
      consumes:
        - application/json
      produces:
        - application/json
      summary: create Organization
      operationId: createOrganization
      description: Creates a new organization
      responses:
        '201':
          description: organization created
          schema:
            $ref: "#/definitions/CreatedResponse"
        '409':
          description: an existing item already exists
          schema:
            $ref: "#/definitions/ConflictResponse"
        '401':
          description: Authorization error
          schema:
            $ref: "#/definitions/ErrorResponse"
        '403':
          description: Authorization error
          schema:
            $ref: "#/definitions/ErrorResponse"
        '500':
          description: unexpected error
          schema:
            $ref: "#/definitions/ErrorResponse"
      parameters:
        - in: body
          name: organization
          description: Organization to be added
          schema:
            $ref: '#/definitions/Organization'
  /organization/{id}:
    get:
      tags:
        - organizationManagement
      consumes:
        - application/json
      produces:
        - application/json
      summary: Get specific organization
      operationId: getOrganization
      description: get organization with given organizationId
      responses:
        '200':
          description: organization returned
          schema:
            $ref: "#/definitions/Organization"
        '404':
          description: organization with organizationId not found
          schema:
            $ref: "#/definitions/MissingResponse"
        '401':
          description: Authorization error
          schema:
            $ref: "#/definitions/ErrorResponse"
        '403':
          description: Authorization error
          schema:
            $ref: "#/definitions/ErrorResponse"
        '500':
          description: unexpected error
          schema:
            $ref: "#/definitions/ErrorResponse"
      parameters:
        - in: path
          name: id
          type: string
          format: uuid
          required: true
          description: Id of organization to be obtained
    put:
      tags:
        - organizationManagement
      consumes:
        - application/json
      produces:
        - application/json
      summary: Update specific organization
      operationId: updateOrganization
      description: Update organization with given organizationId
      responses:
        '200':
          description: updated organization
          schema:
            $ref: "#/definitions/Organization"
        '404':
          description: organization with organizationId not found
          schema:
            $ref: "#/definitions/MissingResponse"
        '401':
          description: Authorization error
          schema:
            $ref: "#/definitions/ErrorResponse"
        '403':
          description: Authorization error
          schema:
            $ref: "#/definitions/ErrorResponse"
        '500':
          description: unexpected error
          schema:
            $ref: "#/definitions/ErrorResponse"
      parameters:
        - in: path
          name: id
          type: string
          format: uuid
          required: true
          description: Id of organization to be obtained
        - in: body
          name: organization
          required: true
          description: updated organization data to be added
          schema:
            $ref: "#/definitions/Organization"
    delete:
      tags:
        - organizationManagement
      consumes:
        - application/json
      produces:
        - application/json
      summary: Delete specific organization
      operationId: deleteOrganization
      description: Delete organization with given organizationId
      responses:
        '200':
          description: deleted organization
          schema:
            $ref: "#/definitions/DeletedResponse"
        '404':
          description: organization with organizationId not found
          schema:
            $ref: "#/definitions/MissingResponse"
        '401':
          description: Authorization error
          schema:
            $ref: "#/definitions/ErrorResponse"
        '403':
          description: Authorization error
          schema:
            $ref: "#/definitions/ErrorResponse"
        '500':
          description: unexpected error
          schema:
            $ref: "#/definitions/ErrorResponse"
      parameters:
        - in: path
          name: id
          type: string
          format: uuid
          required: true
          description: Id of organization to be obtained
  /organization/{id}/user/{userID}:
    put:
      tags:
        - organizationManagement
      consumes:
        - application/json
      produces:
        - application/json
      summary: Adds user to organization
      operationId: addUserToOrganization
      description: Adds user to organization
      responses:
        '200':
          description: user updated
          schema:
            $ref: "#/definitions/OKResponse"
        '404':
          description: organization with organizationId not found
          schema:
            $ref: "#/definitions/MissingResponse"
        '401':
          description: Authorization error
          schema:
            $ref: "#/definitions/ErrorResponse"
        '403':
          description: Authorization error
          schema:
            $ref: "#/definitions/ErrorResponse"
        '500':
          description: unexpected error
          schema:
            $ref: "#/definitions/ErrorResponse"
      parameters:
        - in: path
          name: id
          type: string
          format: uuid
          required: true
          description: Id of organization to be obtained
        - in: path
          name: userID
          type: string
          format: uuid
          required: true
          description: Id of user to be added
    delete:
      tags:
        - organizationManagement
      consumes:
        - application/json
      produces:
        - application/json
      summary: Delete user from organization
      operationId: deleteUserFromOrganization
      description: Delete user from organization
      responses:
        '200':
          description: deleted user from organization
          schema:
            $ref: "#/definitions/DeletedResponse"
        '404':
          description: organization or user not found
          schema:
            $ref: "#/definitions/MissingResponse"
        '401':
          description: Authorization error
          schema:
            $ref: "#/definitions/ErrorResponse"
        '403':
          description: Authorization error
          schema:
            $ref: "#/definitions/ErrorResponse"
        '500':
          description: unexpected error
          schema:
            $ref: "#/definitions/ErrorResponse"
      parameters:
        - in: path
          name: id
          type: string
          format: uuid
          required: true
          description: Id of organization to be obtained
        - in: path
          name: userID
          type: string
          format: uuid
          required: true
          description: Id of user to be removed
...
definitions:
...
  Organization:
    type: object
    properties:
      ID:
        type: string
        format: uuid
        x-go-custom-tag: gorm:"type:uuid;primary_key;unique;default:md5(random()::text || clock_timestamp()::text)::uuid"
      FormalName:
        type: string
        x-go-custom-tag: gorm:"column:formalname"
      RegisteredAddress1:
        type: string
        x-go-custom-tag: gorm:"column:registeredaddress1"
      RegisteredAddress2:
        type: string
        x-go-custom-tag: gorm:"column:registeredaddress2"
      RegisteredAddress3:
        type: string
        x-go-custom-tag: gorm:"column:registeredaddress3"
      RegisteredCountry:
        type: string
        format: country
        x-go-custom-tag: gorm:"column:registeredcountry"
      CreationDate:
        type: string
        format: date-time
        x-go-custom-tag: gorm:"column:registrationdatetime;type:timestamptz;default:now()"
      CreatedBy:
        # userid who created this organization
        type: string
        format: uuid
        x-go-custom-tag: gorm:"column:createdby;type:uuid"
      Website:
        type: string
        format: url
      OrganizationEmailAddress:
        type: string
        format: email
        x-go-custom-tag: gorm:"column:organizationemailaddress"
      PrimaryTelephoneNumber:
        type: string
        format: telephone-number
        x-go-custom-tag: gorm:"column:primarytelephonenumber"
      VATRegistrationNumber:
        type: string
        x-go-custom-tag: gorm:"column:vatregistrationnumber"
      OrganizationStatus:
        type: string
        enum:
        - PENDING_APPROVAL
        - APPROVED
        - DISABLED
        - TERMINATED
        x-go-custom-tag: gorm:"column:organizationstatus"
...
  CreatedResponse:
    type: object
    properties:
      ID:
        type: string
      Link:
        type: string
  ConflictResponse:
    type: object
    properties:
      ID:
        type: string
      Message:
        type: string
  DeletedResponse:
    type: object
    properties:
      ID:
        type: string
      Message:
        type: string
  OKResponse:
    type: object
    properties:
      Message:
        type: string
  ErrorResponse:
    type: object
    properties:
      Message:
        type: string
  MissingResponse:
    type: object
    properties:
      ID:
        type: string
      Message:
        type: string
...